# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/settings": {
            "put": {
                "tags": [
                    "settings"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample setting to test out the API",
                "operationId": "sett",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "key",
                        "required": true
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "value",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/employees": {
            "post": {
                "tags": [
                    "employees"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample employees to test out the API",
                "operationId": "employees",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salary",
                        "in": "query",
                        "description": "salary",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/overtimes": {
            "post": {
                "tags": [
                    "overtimes"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample overtimes to test out the API",
                "operationId": "overtimes",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "employee_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "time_started",
                        "in": "query",
                        "description": "time_started",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "time_ended",
                        "in": "query",
                        "description": "time_ended",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/overtime-pays/calculate": {
            "get": {
                "tags": [
                    "overtime-pays calculate"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample calculate to test out the API",
                "operationId": "calculate",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "month (2022-09)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/testin": {
            "post": {
                "tags": [
                    "testinx"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample testing to test out the API",
                "operationId": "testing",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}